Make the meaning of physics body more flexible
The idea is to be able to treat a whole rocket as a physics body. Since mass is made  up of the mass of its sub components, we cannot have just one number for mass.

The potential counterpoint to this is that on each simulation step, maybe a rocket assembly should just be used to generate a physics body with a total mass.

Then we can do physics calculations entirely separate instead of having an elaborate inheritance hierarchy.