Make it possible to copy rocket, pass tests
The reason we do this is because it turns out that reseting the state of a rocket is very cumbersome. It is error prone. One easily forget some state, which has to be reset. It is easier to always just operate on a copy of the originally created rocket.

Perhaps having a function create a mutable copy would be an alternative, and the rocket you assembly initially is made purely from immutable parts.

This also makes all the tests pass, as we don't get all the failed state resets.