Major refactoring
Removed the booster types. Instead in next step we will add side boosters as an array on the Rocket type. However the boosters will be of type Rocket rather than Booster.

Motivation: The previous data structure did not allow boosters to fire their engines at the same time as the centre stage/core. But that is for next  refactoring step. This step was to retain as much behavior as possible from before and make most tests succeed.

Second motivation was to turn Engine and Tank types into more of a Engine Type and Tanke Type structures. That is making them immutable and keep mutated data such as throttle and propellant left on the rocket object.